% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acummulate_incidence.R
\name{acummulate_incidence}
\alias{acummulate_incidence}
\title{Acummulate incidence from interaction data frame}
\usage{
acummulate_incidence(data, group, incidence = incidence)
}
\arguments{
\item{data}{A \code{data.frame} object with interaction information}

\item{group}{A group column in data \code{data.frame} to acummulate incidence in an interaction \code{data.frame}}

\item{incidence}{The incidence of interaction in the data \code{data.frame}}
}
\value{
A \code{data.frame} with the summarized incidence from an interaction \code{data.frame}. Includes the acummulated incidence for each group provided
}
\description{
Acummulate incidence from interaction data frame
}
\examples{
library(ecointeraction)
 acummulate_incidence(mammalvirus, virus, incidence)
}
