% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batscoronavirus.R
\docType{data}
\name{batscoronavirus}
\alias{batscoronavirus}
\title{batscoronavirus}
\format{
A data frame with 1209 rows and 2 variables:
\describe{
  \item{species}{Bats species}
  \item{incidence}{Number of coronavirus records from dBatVir database for each species of bat }
}
}
\source{
\url{http://http://www.mgc.ac.cn/DBatVir/}
}
\usage{
batscoronavirus
}
\description{
A dataset containing the incidence of coronavirus on bats
The raw-data is from DbatVir dataset. Containsinformation for different coronavirus.
Then data was summarized counting entries for coronavirus taked as raw incidence by bat species
}
\keyword{datasets}
