% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birdsdistance.R
\docType{data}
\name{birdsdistance}
\alias{birdsdistance}
\title{birdsdistance}
\format{
A data frame with 1191 rows and 12 variables:
\describe{
  \item{species}{Bird species}
  \item{geo_distance}{Geographical distance}
  \item{env_distance}{Environmental distance}
  \item{phylo_distance}{Phylogenetic distance}
}
}
\usage{
birdsdistance
}
\description{
A dataset containing the median phylogenetic, geographical and enviromental distances of birds.
The raw-data has the distances for each pair of birds. The data was centered and scaled.
Then was summarized by each species taking the median for each distance.
Environmental information from WorldClim project
Phylogenetic information from Vertlife project
Geografical information from GBIF project
}
\keyword{datasets}
