% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birdsplasmodium.R
\docType{data}
\name{birdsplasmodium}
\alias{birdsplasmodium}
\title{birdsplasmodium}
\format{
A data frame with 1209 rows and 2 variables:
\describe{
  \item{species}{Bird species}
  \item{incidence}{Number of records from Malavi database for each species fo Plasmodium gn.}
}
}
\source{
\url{http://130.235.244.92/Malavi/}
}
\usage{
birdsplasmodium
}
\description{
A dataset containing the incidence of Plasmodium genus on birds
The raw-data is from Malavi dataset. Contains avian malaria information for different Genus.
Then data was summarized counting entries for plasmodium genus taked as raw incidence
}
\keyword{datasets}
