% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birdsplasmodiumrelictum.R
\docType{data}
\name{birdsplasmodiumrelictum}
\alias{birdsplasmodiumrelictum}
\title{birdsplasmodiumrelictum}
\format{
A data frame with 195 rows and 12 variables:
\describe{
  \item{species}{Bird species}
  \item{incidence}{Number of records from Malavi database for each species fo Plasmodium gn.}
}
}
\source{
\url{http://130.235.244.92/Malavi/}
}
\usage{
birdsplasmodiumrelictum
}
\description{
A dataset containing the incidence of Plasmodium relictum on birds
MalAvi lineage code: GRW04, GRW11, LZFUS01 and SGS1
The raw-data is from Malavi dataset. Contains avian malaria information for different lineage.
Then data was summarized counting entries for Plasmodium relictum taked as raw incidence
}
\keyword{datasets}
