% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cummulative_rate.R
\name{cummulative_rate}
\alias{cummulative_rate}
\title{cummulative_rate}
\usage{
cummulative_rate(
  data,
  id = id,
  incidence = incidence,
  cummulativesum = cummulativesum,
  accuracy = 4
)
}
\arguments{
\item{data}{A \code{data.frame} object with interaction information}

\item{id}{Row id of incidence}

\item{incidence}{The incidence of interaction in the data \code{data.frame}}

\item{cummulativesum}{cummulative counted incidences}

\item{accuracy}{of the derivate of the rate}
}
\value{
A \code{data.frame} with the rate of change from fractional cuumlative incidence
and the first and second derivates of the rate
}
\description{
cummulative_rate
}
\examples{
library(ecointeraction)
data <-  acummulate_incidence(mammalvirus, mammal_species, incidence)
cumrateData <- cummulative_rate(data, id, incidence, cummulativesum)
}
