% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_powerlaw.R
\name{get_powerlaw}
\alias{get_powerlaw}
\title{Continuous maximum likelihood estimators for fitting the power-law distribution to data.
This fuction is based on poweRlaw package using the methods described in Clauset et al, 2009.}
\usage{
get_powerlaw(incidence, threads = 1, seed = 123, no_of_sims = 1000)
}
\arguments{
\item{incidence}{The incidence of interaction in the data \code{data.frame}}

\item{threads}{The number of threads used in boostrapping}

\item{seed}{Set the seed for replication}

\item{no_of_sims}{Set the seed number of replications in bootstrap}
}
\value{
A \code{list} with the power law distribution fitted to incidence data and de p-value of the test
}
\description{
Continuous maximum likelihood estimators for fitting the power-law distribution to data.
This fuction is based on poweRlaw package using the methods described in Clauset et al, 2009.
}
\examples{
\dontrun{
#' library(ecointeraction)
data(birdsplasmodiumrelictum)
get_powerlaw(birdsplasmodiumrelictum)
}
}
