% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foo.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe transforming functions}
\arguments{
\item{lhs, rhs}{A visualisation and a function to apply to it}
}
\description{
Like dplyr, ecointeraction also uses the pipe function, \code{\%>\%} to turn
function composition into a series of imperative statements.
}
\examples{
# Instead of cummulative_rate(acummulate_incidence(mammalvirus,
# mammal_species, incidence), id, incidence, cummulativesum)
# you can write
library(ecointeraction)
mammalvirus \%>\%
    acummulate_incidence(mammal_species, incidence) \%>\%
    cummulative_rate(id, incidence, cummulativesum)
}
