% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_incidence_data.R
\name{prep_incidence_data}
\alias{prep_incidence_data}
\title{prep_incidence_data
Function to prepare incidence data with distance data to model. Takes the species
in the incidence data and join with a sample form the distance data.
This allow to get a dataset ready to model with ML framework}
\usage{
prep_incidence_data(incidence, distance, by = "species")
}
\arguments{
\item{incidence}{A \code{data.frame} object with incidence per species}

\item{distance}{A \code{data.frame} with enviromental, geographical and
phylogenetic median distance}

\item{by}{A character vector of variables to join by. The distance table and
the incidence table must have an species column}
}
\value{
A \code{data.frame} with data prepare to model. Includes a sample of median distance
 labeled as unknown class and incidence species labeled as susceptible class
}
\description{
prep_incidence_data
Function to prepare incidence data with distance data to model. Takes the species
in the incidence data and join with a sample form the distance data.
This allow to get a dataset ready to model with ML framework
}
\examples{
library(ecointeraction)
birdsplasmodium \%>\%
    acummulate_incidence(group = species) \%>\%
    cutoff_incidence() \%>\%
    prep_incidence_data(distance  = birdsdistance)
}
